requirements.txt
Django==2.2.5
django-filter==2.2.0
djangorestframework==3.10.3
numpy==1.17.2
pandas==0.25.1
pi==0.1.2
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
